Moved routers into routers directory
