+++ package.json
@@ -48,7 +48,7 @@
     "express": "^4.18.3",
     "express-session": "^1.18.0",
     "knex": "^3.1.0",
-    "pinniped": "^0.1.8",
+    "pinniped": "latest",
     "uuid": "^9.0.1",
     "xss": "^1.0.15"
   },
